fix-windows-build (#34)
* set up workflows

* get rid of all conversion warnings by using size_t at some places

* fixed valgrind workflow

* window threads

* windows timer

* fix include error

* windows abstraction for threads

* activate tests

* improved tests

* ran formatter + C++ test workflow

* format of printing

* windows disable parallel compilation

* debug output windows build

* only build windows

* new attempt windows build release mode

* attempt on MSVC compiler bug

* attempt on MSVC compiler bug

* disable tests

* compile test files separately

* activate tests

* object library test_macros

* commenting out half the tests

* comment out more tests

* comment out more tests

* minor

* add dton and dom prop

* comment out dton and dom prop

* comment out pathlogical

* only run pathlogocail

* comment out some tests in pathological

* add back test with empty arrays

* add back test with empty arrays

* replaced empty arrays with null pointers

* replaced empty arrays with null pointers

* typo

* add more tests

* add even more tests

* add more tests

* added parallel rows test

* comment out test 3 parallel rows

* fixed windows order of parallel rows 3

* debug order of rows test 3

* test dton

* comment out parallel rows test

* adding some parallel row tests

* forgot include

* comment out test 15 parallel rows

* comment out test 16 parallel rows

* add parallel col tests

* add print output

* skip parallel rows and cols for windows

* add mac and ubuntu to cmake workflow

* add postsolve tests for mac and ubuntu

* add postsolve tests to windows to find failing test

* test 6 postsolve windows

* test 6 postsolve

* test 7 postsolve windows column order

* test 9 windows build

* commented out unncessearary code

* introduced PSLP_uint

* Revert "introduced PSLP_uint"

This reverts commit 9dafbef89cba7c1fe004e00c6005b9d6e29b1aff.

* clean up

* clean up cmake

* more clean up cmake

* more clean up

* update version

+++ include/PSLP/PSLP_sol.h
@@ -24,13 +24,15 @@ extern "C"
 {
 #endif
 
+#include <stddef.h>
+
     typedef struct Solution
     {
         double *x;
         double *y;
         double *z;
-        int dim_x;
-        int dim_y;
+        size_t dim_x;
+        size_t dim_y;
     } Solution;
 
 #ifdef __cplusplus
